fixing TabController
