+++ src/Visualisation/Tab.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Visualisation.TabController">
+<AnchorPane fx:id="islandsForm" minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Visualisation.TabController">
     <children>
         <Label layoutX="14.0" layoutY="14.0" text="Starting population" AnchorPane.leftAnchor="5.0" AnchorPane.topAnchor="5.0">
             <font>
